third_party/blink/web_tests/platform/generic/inspector-protocol/runtime/runtime-evaluate-user-gesture-expected.txt

Tests that Runtime.evaluate works with userGesture flag.

Running test: testInitialUserActivation
{
    result : {
        type : boolean
        value : false
    }
}

Running test: testActiveWithoutUserGesture
{
    result : {
        type : boolean
        value : false
    }
}

Running test: testActiveWithUserGesture
{
    result : {
        type : boolean
        value : true
    }
}
